# Some optimizations.