--- a/include/ada/mimesniff/util.h
+++ b/include/ada/mimesniff/util.h
@@ -12,8 +12,6 @@ constexpr inline void trim_trailing_http_whitespace(std::string_view& input);
 
 constexpr inline bool is_http_whitespace(char c);
 
-constexpr static inline bool is_http_token(char c);
-
 /**
  * Assuming that the view is UTF-8 encoded, we return true if its characters are
  * U+0009 TAB, a code point in the range U+0020 SPACE to U+007E (~) or in the
@@ -21,8 +19,16 @@ constexpr static inline bool is_http_token(char c);
  * @see https://mimesniff.spec.whatwg.org/#http-token-code-point
  */
 constexpr inline bool contains_only_http_tokens(std::string_view view);
+/**
+ * Assuming that the view is UTF-8 encoded, we return a byte value with the most
+ * significant bit set to 0 if its characters are
+ * U+0009 TAB, a code point in the range U+0020 SPACE to U+007E (~) or in the
+ * range U+0080 through U+00FF (ÿ), inclusive. The 3rd bit is set to 1 if an
+ * uppercase ASCII letter was found.
+ * @see https://mimesniff.spec.whatwg.org/#http-token-code-point
+ */
+constexpr inline uint8_t http_tokens_map(std::string_view view);
 
-constexpr static inline bool is_http_quoted_string_token(char c);
 
 /**
  * @see https://mimesniff.spec.whatwg.org/#http-quoted-string-token-code-point
@@ -39,5 +45,10 @@ inline std::string collect_http_quoted_string(std::string_view input,
  */
 constexpr bool to_lower_ascii(char* input, size_t length) noexcept;
 
+/**
+ * Lowers the ASCII letters in the string in-place. This function is optimized
+ * for short inputs (<= 8 characters).
+ */
+constexpr void to_lower_ascii_short(char* input, size_t length) noexcept;
 }  // namespace ada::mimesniff
 #endif