On our benchmarks, this PR improves our speed by about 10%. It is not much but the changes are tiny.



Intel Icelake, GCC 12...

Before ...

```
BasicBench_mean       155538 ns       155317 ns            3 GHz=3.19359 cycle/byte=49.6193 cycles/mime=559.33 instructions/byte=159.229 instructions/cycle=3.209 instructions/mime=1.79489k instructions/ns=10.2483 mime/s=5.67226M/s ns/mime=175.141 speed=63.9401M/s time/byte=15.6396ns time/mime=176.297ns
```

After this PR...

```
BasicBench_mean       138132 ns       137944 ns            3 GHz=3.19368 cycle/byte=44.3077 cycles/mime=499.455 instructions/byte=149.952 instructions/cycle=3.38434 instructions/mime=1.69033k instructions/ns=10.8085 mime/s=6.38667M/s ns/mime=156.389 speed=71.9932M/s time/byte=13.8902ns time/mime=156.576ns
```

For some context, the cost of parsing a mime is about 70% of the cost of parsing a URL. On a per byte basis, it takes 150 instructions per byte to parse a mime, and 25 instructions per byte to parse a URL.

The benchmarks are not directly comparable, but mime parsing is not cheap (yet).